Added a flow_calc function to use the efficiency curve, small bug in the DNI data import, bug  in lim_inf_DNI
The lim_inf _DNI was hada a mistaken (-) sign and has been fixed. Added the DNI[i]>0 condition to the main for loop since some lim_inf_DNI were negative due the ambient temperatures.

The column numbers in the import of the DNI data were wrong in the CIMAV case and have been fixed.

Created a new "flow_calc" for obtaining the same results as the old one but from the efficiency curve (the change in the fluid caudal hasn't been taken into account). Modified every call to this function from "operationSimple" for using the new equation in the CIMAV sender case.

Everything seems to be working fine.